code/python/requirements.txt
Flask==2.0.1
google-cloud-pubsub==2.8.0
google-cloud-secret-manager==2.8.0
#google-cloud-sql==2.8.0
cloud-sql-python-connector==1.7.0
